d3d12: Fix d3d12_video_enc.cpp(4794,33): Error C4244: initializing: conversion from uint64_t to SIZE_T, possible loss of data
Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/37982>

+++ src/gallium/drivers/d3d12/d3d12_video_enc.cpp
@@ -4791,7 +4791,7 @@ d3d12_video_encoder_extract_encode_metadata(
 #endif
 
    // Map metadata buffer using native D3D12 API
-   D3D12_RANGE readRange = { 0, resourceMetadataSize };
+   D3D12_RANGE readRange = { 0, static_cast<SIZE_T>(resourceMetadataSize) };
    void *pMetadataBufferSrc;
    HRESULT hr = pResolvedMetadataBuffer->Map(0, &readRange, &pMetadataBufferSrc);
    if (FAILED(hr)) {
